bug 1120193, update mercurial to 3.2.4.
The actual mercurial dependencies are in elmo's pushes.utils. That
didn't require any code changes for our use of mercurial apis, though.

The biggest part of the patch here is to use version-control-tools for
the staging environment instead of the separate hgcustom repos. This
brings our staging hg server in line with what's currently running in
production on hg.mozilla.org.