Refactor create-test-env, split hg and automation virtualenvs
Split the virtualenvs, because the version of hg for the hg
server depends on version-control-tools, and we may want to support
a different one for automation.

I also made the setup script a lot simpler, by avoiding to
activate the actual virtualenvs, and just hard-code paths to
pip and hg to absolute paths in the respective bins.