Allow for empty repositories and ones without pushes but content.
When a repository doesn't have any pushes yet, either because
it's empty, or because it got clone upstream (happens for ESR),
make sure we have a local clone in a10n/elmo.
The fix includes allow the list of pushes to be empty when
sending data from get-pushes to hg worker.
There are a few minor adjustments in get-pushes to make that
happen only once on startup. Repositories staying without
pushes for a longer time should probably be archived.

When repos are non-empty, we find the remote heads by
asking the hg server directly, and add those and their parents to
the elmo db.